Remove test file for submission and optimize some functions

+++ ft_putnbr_fd.c
@@ -6,7 +6,7 @@
 /*   By: bvan-den <bvan-den@codam.nl>                 +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2022/10/23 22:52:04 by bvan-den      #+#    #+#                 */
-/*   Updated: 2022/10/23 22:52:05 by bvan-den      ########   odam.nl         */
+/*   Updated: 2022/10/29 19:23:20 by bvan-den      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,10 +20,10 @@ void	ft_putnbr_fd(int n, int fd)
 	if (n < 0)
 	{
 		ft_putchar_fd('-', fd);
-		num = n * -1;
+		num = (unsigned int)(n * -1);
 	}
 	else
-		num = n;
+		num = (unsigned int)n;
 	if (num > 9)
 		ft_putnbr_fd(num / 10, fd);
 	ft_putchar_fd(num % 10 + '0', fd);
