+++ ft_striteri.c
@@ -6,7 +6,7 @@
 /*   By: bvan-den <bvan-den@codam.nl>                 +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2022/10/23 22:24:48 by bvan-den      #+#    #+#                 */
-/*   Updated: 2022/10/25 21:33:50 by bvan-den      ########   odam.nl         */
+/*   Updated: 2022/10/29 19:23:30 by bvan-den      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	ft_striteri(char *s, void (*f)(unsigned int, char *))
 {
-	size_t	i;
+	unsigned int	i;
 
 	i = 0;
 	while (s[i])
